Refactor maze generation and agent movement logic; add Laberinto class for improved structure
